# test: add performance test infrastructure and mock harness