--- a/docs/phase8/agent-core-mvp-status-23may.md
+++ b/docs/phase8/agent-core-mvp-status-23may.md
@@ -1,0 +1,104 @@
+# Agent-core MVP Status Report
+
+**Date**: 23 May 2025  
+**Time**: 21:00 UTC  
+**Phase**: Ready for Performance Validation
+
+## ✅ Completed Today
+
+### 1. PR Merges
+- **PR #345** - Test & performance harness ✅
+  - Fixed CI issues (torch version constraints)
+  - Added env var support to perf harness
+  - Merged at 20:43:59Z
+  
+- **PR #347** - Documentation update ✅
+  - Updated integration surface
+  - Marked Test & Perf harness as **Merged**
+  - Just merged moments ago
+
+### 2. Performance Test Infrastructure
+- ✅ Mock performance test generator (`perf/mock_performance_results.py`)
+- ✅ Comprehensive test plan (`perf/PERFORMANCE_TEST_PLAN.md`)
+- ✅ 40-query test set (`perf/queryset.txt`)
+- ✅ k6 advanced test script (`perf/k6-script.js`)
+- ✅ Release tagging script (`scripts/tag-agent-core-release.sh`)
+- ✅ CI workflow for mock tests (`.github/workflows/mock-perf-gate.yml`)
+
+### 3. Mock Performance Results
+```
+📊 Performance Test Results
+========================================
+Total requests: 600
+Successful: 597
+Error rate: 0.50%
+
+Latency percentiles:
+  p50: 108.96ms
+  p95: 215.90ms ✅
+  p99: 292.80ms
+
+✅ PASS
+```
+
+## 📋 Tomorrow's Tasks (24 May)
+
+| Time | Task | Owner |
+|------|------|-------|
+| 09:00 | Spin up real stack with Go/Docker | You |
+| 09:30 | Run actual performance harness | You |
+| 10:00 | Confirm p95 < 300ms, err < 1% | You |
+| 10:15 | Tag v0.9.0 release | You |
+| 10:30 | Notify BizDev in Slack | You |
+
+## 🚀 Commands Ready to Execute
+
+### 1. Build and Start Services
+```bash
+go build -o alfred ./cmd/alfred/main.go
+alfred up -d pg redis minio server
+```
+
+### 2. Seed Vector Store
+```bash
+export POSTGRES_DSN="postgresql://alfred:alfred@localhost:5432/alfred?sslmode=disable"
+export OPENAI_API_KEY=sk-...
+alfred ingest ./docs/**/*.md --batch 64
+```
+
+### 3. Run Performance Test
+```bash
+export TARGET_URL=http://localhost:8080/v1/query
+export QPS=10
+export DURATION=60
+python perf/harness_scaffold.py | tee /tmp/harness.out
+```
+
+### 4. Tag Release (if tests pass)
+```bash
+./scripts/tag-agent-core-release.sh
+```
+
+## 📊 Integration Status
+
+| Component | Status | PR |
+|-----------|--------|-----|
+| Vector schema migration | **Merged** | #336 |
+| Ingest CLI & indexer | **Merged** | #339 |
+| Retrieval API & RAG loop | **Merged** | #343 |
+| Test & Perf harness | **Merged** | #345 |
+| Documentation | **Updated** | #347 |
+
+## 🎯 Definition of Done
+
+- [x] All implementation PRs merged
+- [x] CI issues resolved
+- [x] Documentation updated
+- [x] Mock performance tests passing
+- [ ] Real performance tests passing (tomorrow)
+- [ ] v0.9.0 tagged and pushed
+- [ ] BizDev notified
+
+---
+
+**Status**: Implementation complete, awaiting real performance validation tomorrow.