--- a/perf/queryset.txt
+++ b/perf/queryset.txt
@@ -1,0 +1,40 @@
+How do I upgrade the Helm chart?
+What is the default p95 latency target?
+How does the EmbeddingRepo interface work?
+What are the key metrics exported by the retrieval endpoint?
+Explain the vector schema design for PostgreSQL
+What is the purpose of the IVFFLAT index?
+How do I configure OpenAI API keys?
+What are the required environment variables for the retrieval server?
+How does the agent-core MVP handle citations?
+What is the maximum query length allowed?
+How many results does the retrieval endpoint return by default?
+What is the timeout for retrieval requests?
+How do I monitor the health of the retrieval service?
+What Prometheus metrics are available?
+How do I scale the retrieval service horizontally?
+What is the recommended batch size for document ingestion?
+How does the platform ensure p95 latency stays under 300ms?
+What are the PostgreSQL requirements for pgvector?
+How do I troubleshoot high latency issues?
+What is the purpose of the GIN index on the ts column?
+How do I add new documents to the vector store?
+What is the embedding dimension for text-embedding-3-small?
+How do I configure the top_k parameter?
+What is the maximum value for top_k?
+How does the system handle concurrent requests?
+What is the recommended connection pool size?
+How do I backup the vector store?
+What are the memory requirements for the retrieval server?
+How do I enable debug logging?
+What is the format of the citation references?
+How do I update existing documents in the vector store?
+What happens when the OpenAI API rate limit is exceeded?
+How do I configure retry logic for failed embeddings?
+What is the purpose of the metadata JSONB column?
+How do I filter search results by metadata?
+What is the recommended deployment strategy?
+How do I perform a rolling update?
+What are the security best practices for API keys?
+How do I enable TLS for the retrieval endpoint?
+What is the expected startup time for the service?